requirements.txt
streamlit
openai
trafilatura
python-dotenv
